Move TimerFactory to its own module
