+++ dds/src/implementation/utils/mod.rs
@@ -1,3 +1,3 @@
-pub mod shared_object;
-pub mod timer;
 pub mod condvar;
+pub mod shared_object;
+pub mod timer_factory;
